added timer
timer is now added however there is a problem when going through items. If click next and simultaneously clicking View Pager, the view gets stuck in between the current and next page.

+++ app/src/main/java/com/eltech/elhealth/Workouts/Workout.java
@@ -8,7 +8,7 @@ public abstract class Workout{
     public abstract int getTitle();
     public abstract String getReps(int level);
     public abstract boolean hasTimer();
-    public abstract int getTimer(int timer,int level);
+    public abstract long getTimer(long timer,int level);
     public abstract @DrawableRes int getImage();
     public abstract boolean affectsKnee();
     public abstract boolean requiresEquipments();
